schemesmart-ai/backend/requirements.txt
fastapi
uvicorn
supabase
python-multipart
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
scikit-learn
pydantic
groq
python-dotenv
gTTS
